requirements.txt
mysql-connector
mysql-connector-python
flask
flask_sqlalchemy
Flask-MySQLdb
requests
pandas
numpy